64 bit compile changes - not so many casts.

+++ freewrl/CFuncs/PluginSocket.c
@@ -1,5 +1,5 @@
 /*
- * $Id: PluginSocket.c,v 1.15 2006/02/08 21:25:17 crc_canada Exp $
+ * $Id: PluginSocket.c,v 1.16 2006/03/02 20:44:09 crc_canada Exp $
  *
  * Common functions used by Mozilla and Netscape plugins...(maybe
  * PluginGlue too?)
@@ -93,7 +93,7 @@ int waitForData(int sock) {
 
 char *
 requestUrlfromPlugin(int sockDesc,
-		   unsigned int plugin_instance,
+		   uintptr_t plugin_instance,
 		   const char *url)
 {
 	size_t len = 0, ulen = 0, bytes = 0;
@@ -190,7 +190,7 @@ requestUrlfromPlugin(int sockDesc,
  * clicked and it is an HTML page */
 
 void requestNewWindowfromPlugin(int sockDesc,
-		   unsigned int plugin_instance,
+		   uintptr_t plugin_instance,
 		   const char *url)
 {
 	size_t len = 0, ulen = 0, bytes = 0;
